Update Reddit_AddIn.bas
So, it seems to me that the previous version was looking for the formatting of the wrong row in the referenced range. Also, I've added conditional compiler directives so that the add-in can be run on both the Mac and Windows operating systems. Um, did some dusting and gave myself credit!